4.2.0 feat(ss-core): fastjson升级到1.2.76 feat(ss-netty-*): 重构 feat(ss-processor): 删除ss-core的exclusion

+++ ss-netty-parent/ss-netty-client/src/main/java/com/mxny/ss/netty/client/cache/ClientCache.java
@@ -17,5 +17,5 @@ public class ClientCache {
     //连接器
     public static DefaultCommonClientConnector clientConnector;
 
-    public final static ConcurrentMap<Long, MessageNonAck> messagesNonAcks = new ConcurrentHashMap<Long, MessageNonAck>();
+    public final static ConcurrentMap<Integer, MessageNonAck> messagesNonAcks = new ConcurrentHashMap<Integer, MessageNonAck>();
 }
